Allow top-level functions to call themselves recursively
This requires us to be able to get top-level functions. To do this a
new instruction was introduced - GETA - which gets a local with a value
which is counted from the bottom of the global locals stack (rather
than the chunk's part of the local stack in the case of GET).