Stop infinite loop when checking functions with wrong arguments
Expressions like `-true` would fail to terminate previous to this
change due to repeatedly expanding the same function constructor and
pruning the same resultant constraints. This was making progress even
though the final result was always the same.

To fix, just make sure that when expanding constructors the arguments
can actually be unified.