Introduce scoping to variables
This is a precursor to let..in. We introduce a new datastructure which
is a list of maps from identifier to type variable. Whenever a scope is
introduced - e.g. at the start of a function (for the parameters,
implemented), or during a let..in (not implemented) - we add a new map.
This then allows shadowing to happen.